fixed main code

+++ src/main/java/service/operation/PurchaseOperation.java
@@ -6,6 +6,11 @@ import model.FruitTransaction;
 public class PurchaseOperation implements OperationHandler {
     @Override
     public void performOperation(FruitTransaction fruitTransaction) {
-        Storage.remove(fruitTransaction.getFruit(), fruitTransaction.getQuantity());
+        if (Storage.getAmount(fruitTransaction.getFruit()) >= fruitTransaction.getQuantity()) {
+            Storage.remove(fruitTransaction.getFruit(), fruitTransaction.getQuantity());
+        } else {
+            throw new IllegalArgumentException("Not enough "
+                    + fruitTransaction.getFruit() + " in the store.");
+        }
     }
 }
